backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
stripe>=7.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
